[cli] rename trusted_commands.rs to trusted_cli.rs
