+++ cli/src/trusted_cli.rs
